train_requirements.txt
# Core dependencies for COMET training script
unbabel-comet==2.2.2
torch>=1.11.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Optional but recommended for better performance
transformers>=4.15.0
sentencepiece>=0.1.96
protobuf==3.20.*

# For data handling and utilities
tqdm>=4.62.0